좌표계 수정

+++ 02_10motor/inc/Homo_Coordinate.h
@@ -40,9 +40,9 @@
 
 /*********************************************************************************************************************/
 /*-------------------------------------------------Global variables--------------------------------------------------*/
-extern float32 cam_points[4][2];
-extern float32 world_points[4][2];
-extern float32 H[3][3];
+extern double cam_points[4][2];
+extern double world_points[4][2];
+extern double H[3][3];
 extern int data_ready_flag;
 extern int transform_finished;
 /*********************************************************************************************************************/
@@ -58,7 +58,7 @@ extern int transform_finished;
 /*********************************************************************************************************************/
 /*------------------------------------------------Function Prototypes------------------------------------------------*/
 void TransformData_Init(void);
-void transform_points(float32 H[3][3], float32 cam_points[4][2], float32 world_points[4][2]);
+void transform_points(double H[3][3], double cam_points[4][2], double world_points[4][2]);
 void InitCampoints(void);
 void InitWorldpoints(void);
 /*********************************************************************************************************************/
